Fix event timestamps not aligning to actual real time
In the editor, the input system uses an internal time offset which is adjusted whenever play mode is entered or exited. This offset is used to make the 0 point of input timestamps match the time at which the last play mode transition happened.

The current method of grabbing input time bypasses this offset by using reflection to access an internal property. Without this offset, timestamps will never be in sync while in the editor, and will continually drift more and more each time play mode is entered.

The publicly-accessible `InputState.currentTime` property accounts for this offset, and so do the `InputSystem.Queue*Event` family of functions. Using that instead fixes this issue.